Refactor Styling
- Move all MUI styling and theme into separate files
- Split up Event Show
- Consistent file/folder structures
- Remove unnecessary login/signup code
