+++ client/src/components/eventForm/Delete.js
@@ -1,44 +1,10 @@
 import { Button } from '@material-ui/core';
 import React, { useState } from 'react'
 import { Redirect } from "react-router-dom"
-import { makeStyles, createMuiTheme, ThemeProvider } from '@material-ui/core/styles'
+import { ThemeProvider } from '@material-ui/core/styles'
 import { Grid, Typography } from '@material-ui/core'
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: '#ED1A7A',
-      main: '#ED1A7A',
-      dark: '#ED1A7A',
-      contrastText: '#000',
-    },
-    secondary: {
-      light: '#49AEB9',
-      main: '#49AEB9',
-      dark: '#49AEB9',
-      contrastText: '#fff',
-    }
-  }
-});
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    '& .MuiTextField-root': {
-      margin: theme.spacing(2),
-    },
-    flexGrow: 1,
-    marginTop: theme.spacing(3),
-    backgroundColor: '#EEEEEE',
-    border: 'solid 5px #ED1A7A',
-    borderRadius: theme.spacing(4),
-    textAlign: 'center',
-    padding: theme.spacing(3),
-    color: '#33485E',
-  },
-  button: {
-    color: 'white'
-  }
-}));
+import theme from '../mui/theme.js'
+import useStyles from './styling.js' 
 
 const Delete = props => {
   const classes = useStyles();
