+++ client/src/components/eventShow/Logic.js
@@ -1,110 +1,10 @@
 import React, { useState, useEffect } from "react"
 import { Link, withRouter } from "react-router-dom"
-import { makeStyles, createMuiTheme, ThemeProvider, responsiveFontSizes } from '@material-ui/core/styles'
-import { Grid, Divider, Typography, Button } from '@material-ui/core'
-import Carousel from 'react-material-ui-carousel'
+import { Grid, Divider, Typography } from '@material-ui/core'
 import ReactPlayer from 'react-player'
 import { Delete, Edit } from '@material-ui/icons/';
-
-let theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: '#ED1A7A',
-      main: '#ED1A7A',
-      dark: '#ED1A7A',
-      contrastText: '#000',
-    },
-    secondary: {
-      light: '#49AEB9',
-      main: '#49AEB9',
-      dark: '#49AEB9',
-      contrastText: '#fff',
-    }
-  }
-});
-theme = responsiveFontSizes(theme)
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-    margin: theme.spacing(5),
-    backgroundColor: '#EEEEEE',
-    border: 'solid 5px #ED1A7A',
-    borderRadius: theme.spacing(4),
-    padding: theme.spacing(1),
-    color: '#33485E',
-    width: '90%',
-    display: 'flex',
-    textAlign: 'center',
-    flexWrap: 'wrap',
-    justifyContent: 'space-around',
-    overflow: 'hidden',
-    backgroundColor: '#EEEEEE',
-  },
-  divider: {
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(1)
-  },
-  title: {
-    color: '#ED1A7A',
-    fontWeight: 'bold',
-  },
-  date: {
-    color: '#ED1A7A',
-    fontWeight: 'bold',
-  },
-  time: {
-    color: '#49AEB9',
-  },
-  avatar: {
-    maxWidth: 15,
-    maxHeight: 15,
-    borderRadius: 10,
-  },
-  username: {
-  },
-  headers: {
-    color: '#33485E',
-    fontWeight: 'bold',
-  },
-  eventDescription: {
-    color: '#33485E',
-    textAlign: 'left',
-    margin: '1vw 1vw 0 1vw'
-  },
-  gameSummary: {
-    color: '#49AEB9',
-    textAlign: 'left',
-    margin: '1vw 1vw 0 1vw'
-  },
-  platformImage: {
-    maxHeight: 50,
-    maxWidth: 50
-  },
-  platforms: {
-    paddingTop: theme.spacing(2),
-    color: '#969696',
-  },
-  carouselImage: {
-    height: 250,
-    backgroundRepeat: "no-repeat",
-    backgroundPosition: "center",
-    backgroundSize: "cover",
-    border: 'solid 5px #33485E',
-    borderRadius: theme.spacing(3)
-  },
-  carouselVideo: {
-    padding: theme.spacing(1),
-    marginLeft: 'auto',
-    marginRight: 'auto'
-  },
-  coverArt: {
-    padding: theme.spacing(2)
-  },
-  buttons: {
-    margin: theme.spacing(1)
-  }
-}));
+import Show from './Show.js'
+import useStyles from './styling.js'
 
 const EventShow = props => {
   const classes = useStyles();
@@ -300,78 +200,20 @@ const EventShow = props => {
   }, [game])
 
   return (
-    <ThemeProvider theme={theme}>
-      <Grid container
-        direction="row"
-        justify="space-evenly"
-        alignItems="center"
-        className={classes.root}>
-        <Grid item xs={12} sm={1}>
-          <div className={classes.buttons}>{editButton}</div>
-          <div className={classes.buttons}>{deleteButton}</div>
-        </Grid>
-        <Grid item xs={8} sm={8} >
-          <Typography variant='h2' className={classes.title}>{event.name}</Typography>
-        </Grid>
-        <Grid item xs={4} sm={3}>
-          {event.startDate && getDate(event.startDate, event.endDate)}
-        </Grid>
-        <Grid item xs={12}>
-          <Divider className={classes.divider} light />
-        </Grid>
-        <Grid item xs={12} sm={3}>
-          <a href={`${game.url}`} target="_blank">
-            <img src={coverArt} className={classes.coverArt} />
-          </a>
-        </Grid>
-        <Grid item xs={9} sm={7}>
-          <Typography variant='body1' className={classes.eventDescription}>{event.description}</Typography>
-          <Typography variant='body2' className={classes.gameSummary}>{game && game.summary}</Typography>
-          <Grid container
-            direction="row"
-            justify="center"
-            alignItems="center"
-            spacing={1}
-            className={classes.platforms}>
-            {game && game.platforms && getPlatforms(game.platforms)}
-          </Grid>
-        </Grid>
-        <Grid item xs={3} sm={2}>
-          <Typography variant='h6' className={classes.headers}>Host:</Typography>
-          {event.user && getEventCreator(event)}
-          <Divider className={classes.divider} light />
-          <Typography variant='h6' className={classes.headers}>Location:</Typography>
-          <div><Typography variant='caption'>{event.location && event.location}</Typography></div>
-          <div><Typography variant='caption'> {event.meetUrl && getMeetUrl(event.meetUrl)}</Typography></div>
-          <Divider className={classes.divider} light />
-          <Typography variant='h6' className={classes.headers}>Attending:</Typography>
-          {event.userInterests && getInterestedUsers(event, 'attending')}
-          <Divider className={classes.divider} light />
-          <Typography variant='h6' className={classes.headers}>Interested:</Typography>
-          {event.userInterests && getInterestedUsers(event, 'interested')}
-        </Grid>
-        <Grid item xs={12}>
-          <Divider className={classes.divider} light />
-        </Grid>
-        <Grid item xs={12} md={5} className={classes.media}>
-          <Carousel
-            swipe
-            next={() => { }}
-            prev={() => { }}>
-            {carouselImages}
-          </Carousel>
-        </Grid>
-        <Grid item xs={12} md={5} className={classes.media}>
-          <Carousel
-            autoPlay={false}
-            swipe
-            next={() => { }}
-            prev={() => { }}>
-            {carouselVideos}
-          </Carousel>
-        </Grid>
-      </Grid>
-    </ThemeProvider >
+    <Show 
+    event={event}
+    game={game}
+    carouselImages={carouselImages}
+    carouselVideos={carouselVideos}
+    getDate={getDate}
+    getEventCreator={getEventCreator}
+    getInterestedUsers={getInterestedUsers}
+    getPlatforms={getPlatforms}
+    coverArt={coverArt}
+    getMeetUrl={getMeetUrl}
+    editButton={editButton}
+    deleteButton={deleteButton}
+    />
   )
 }
