+++ client/src/components/eventTile/Tile.js
@@ -1,73 +1,15 @@
 import React from 'react';
 
 import cx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
 import { Link } from "react-router-dom";
 import ArrowForwardIos from '@material-ui/icons/ArrowForwardIos';
 import { usePushingGutterStyles } from '@mui-treasury/styles/gutter/pushing';
 import { useLabelIconStyles } from '@mui-treasury/styles/icon/label';
 import { useRowFlexStyles } from '@mui-treasury/styles/flex/row';
 import { Grid, Avatar, Card, CardContent, CardMedia, Divider } from '@material-ui/core';
+import useStyles from './styling.js'
 
-const useStyles = makeStyles(({ spacing, palette }) => ({
-  card: {
-    display: 'flex',
-    padding: spacing(2),
-    borderRadius: 16,
-  },
-  media: {
-    minWidth: '25%',
-    maxWidth: '25%',
-    flexShrink: 0,
-    backgroundColor: palette.grey[200],
-    borderRadius: 12,
-    boxShadow: '0 2px 8px 0 #c1c9d7, 0 -2px 8px 0 #cce1e9',
-  },
-  rating: {
-    verticalAlign: 'text-top',
-  },
-  content: {
-    padding: spacing(0, 2, 0, 2),
-  },
-  heading: {
-    fontSize: '2.5vw',
-    fontWeight: 'bold',
-    letterSpacing: '0.5px',
-    color: 'rgb(235,35,122)',
-    marginBottom: 0,
-    marginRight: spacing(1.5),
-    display: 'inline-block',
-  },
-  date: {
-    fontSize: 15,
-    fontWeight: 'light',
-    letterSpacing: '0.5px',
-    color: 'rgb(51,72,94)',
-    marginBottom: 0,
-    marginRight: spacing(1.5),
-    display: 'inline-block',
-  },
-  body: {
-    fontSize: 14,
-    color: palette.grey[500],
-  },
-  divider: {
-    margin: spacing(1, 0),
-  },
-  textFooter: {
-    fontSize: 14,
-  },
-  icon: {
-    fontSize: '1.2rem',
-    verticalAlign: 'bottom',
-  },
-  avatar: {
-    maxWidth: 30,
-    maxHeight: 30
-  }
-}));
-
-const EventTileStyle = ({ event, interested, attending }) => {
+const Tile = ({ event, interested, attending }) => {
   const styles = useStyles();
   const gutterStyles = usePushingGutterStyles({ space: 1.5 });
   const labelStyles = useLabelIconStyles({ linked: true });
@@ -139,4 +81,4 @@ const EventTileStyle = ({ event, interested, attending }) => {
   );
 };
 
-export default EventTileStyle
+export default Tile
