+++ client/src/components/eventTile/styling.js
@@ -1,0 +1,61 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(({ spacing, palette }) => ({
+  card: {
+    display: 'flex',
+    padding: spacing(2),
+    borderRadius: 16,
+  },
+  media: {
+    minWidth: '25%',
+    maxWidth: '25%',
+    flexShrink: 0,
+    backgroundColor: palette.grey[200],
+    borderRadius: 12,
+    boxShadow: '0 2px 8px 0 #c1c9d7, 0 -2px 8px 0 #cce1e9',
+  },
+  rating: {
+    verticalAlign: 'text-top',
+  },
+  content: {
+    padding: spacing(0, 2, 0, 2),
+  },
+  heading: {
+    fontSize: '2.5vw',
+    fontWeight: 'bold',
+    letterSpacing: '0.5px',
+    color: 'rgb(235,35,122)',
+    marginBottom: 0,
+    marginRight: spacing(1.5),
+    display: 'inline-block',
+  },
+  date: {
+    fontSize: 15,
+    fontWeight: 'light',
+    letterSpacing: '0.5px',
+    color: 'rgb(51,72,94)',
+    marginBottom: 0,
+    marginRight: spacing(1.5),
+    display: 'inline-block',
+  },
+  body: {
+    fontSize: 14,
+    color: palette.grey[500],
+  },
+  divider: {
+    margin: spacing(1, 0),
+  },
+  textFooter: {
+    fontSize: 14,
+  },
+  icon: {
+    fontSize: '1.2rem',
+    verticalAlign: 'bottom',
+  },
+  avatar: {
+    maxWidth: 30,
+    maxHeight: 30
+  }
+}));
+
+export default useStyles
